% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_functions.R
\name{A.excursion}
\alias{A.excursion}
\title{Excursion area}
\usage{
A.excursion(X, A, n = 10000, level = "greater", subI = NULL, plot = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence (t)
and the second the values of the time series \eqn{(X(t))}.}

\item{A}{Constant level as a numeric value. Checks whether \eqn{X(t)} greater or lower than \code{A}.}

\item{n}{Number of points a time interval to be split into. Default set to 10000.}

\item{level}{A vector of character strings which specifies whether the excursion
area is required for \code{X} \code{"greater"} or \code{"lower"} than \code{A}. Default set to \code{"greater"}.}

\item{subI}{Time sub-interval is a vector where the lower bound is
the first element and upper bound is the second. Optional: If provided the excursion area
of the sub-interval is returned, otherwise the whole time sequence is considered.}

\item{plot}{Logical: If \code{TRUE}, the time series, constant level and excursion area are plotted.}
}
\value{
Excursion area. If \code{plot=TRUE}, the time series, constant level and excursion area
are plotted.
}
\description{
Computes the excursion area where \eqn{X(t)} is greater or lower than the
constant level \code{A} for the provided time interval or a its sub-interval.
}
\examples{
t <- seq(0,1,length=1000)
TS <- data.frame("t"=t,"X(t)"=rnorm(1000))
A.excursion(TS,0.8,level='lower',subI=c(0.5,0.8),plot=TRUE)

}
\seealso{
\code{\link{sojourn}}
}
