% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Bm}
\alias{Bm}
\title{Simulation of Brownian motion}
\usage{
Bm(x1 = 0, t1 = 0, T = 1, N = 1000, plot = TRUE)
}
\arguments{
\item{x1}{Value of the process at the initial time point.}

\item{t1}{Initial time point.}

\item{T}{Terminal time point.}

\item{N}{Number of time steps the interval \eqn{[t1,T]} is split into.
Default set to 1000.}

\item{plot}{Logical: If \code{TRUE}, the realisation of the Brownian
motion is plotted.}
}
\value{
A data frame where the first column is \code{t} and second
column is simulated values of the realisation of Brownian motion.
}
\description{
This function simulates a realisation of the Brownian motion over
time interval \eqn{[t1,T]} with \code{N} increments and initial value \code{x1}.
}
\examples{
Bm(x1=0,t1=0,T=2,N=1000,plot=TRUE)
}
\seealso{
\code{\link{GHBMP}}
}
