% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{plot.hc_hurst}
\alias{plot.hc_hurst}
\title{Plot smoothed Hurst functions in each cluster with cluster centers}
\usage{
\method{plot}{hc_hurst}(x, type = "estimates", ...)
}
\arguments{
\item{x}{Return from \code{\link{hclust_hurst}}.}

\item{type}{The type of plot required:
\describe{
\item{\code{"estimates"}}{Only the smoothed Hurst functions in each cluster.}
\item{\code{"centers"}}{Only the cluster centers. Center denotes average of each smoothed Hurst function in the cluster}
\item{\code{"ec"}}{Both \code{"estimates"} and \code{"centers"}.}
}}

\item{...}{Unused arguments}
}
\value{
A ggplot object which plots the relevant \code{type}.
}
\description{
Creates a plot of the smoothed Hurst functions of each multifractional process in each cluster with cluster centers using the return from
\code{\link{hclust_hurst}}. Options to plot only estimates, only centers or both are available.
}
\examples{
-
#See Examples in hclust_hurst().
}
\seealso{
\code{\link{hclust_hurst}}
}
