% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_est.R
\name{plot_ts_est}
\alias{plot_ts_est}
\title{Plot the estimated Hurst functions and local fractal dimension estimates
for a user provided time series}
\usage{
plot_ts_est(
  X,
  Raw_Est_H = TRUE,
  Smooth_Est_H = TRUE,
  LFD_Est = TRUE,
  LFD_Smooth_Est = TRUE,
  N = 100,
  Q = 2,
  L = 2
)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence \eqn{t}
and the second one is the values of the time series \eqn{X(t)}.
For accurate estimated Hurst functions, \code{X} should be of at least 500 data points.}

\item{Raw_Est_H}{Logical: If \code{TRUE}, the Hurst function estimated by using \code{\link{Hurst}} is plotted.}

\item{Smooth_Est_H}{Logical: If \code{TRUE}, the smoothed estimated Hurst function is plotted.
The estimated Hurst function is smoothed using the loess method.}

\item{LFD_Est}{Logical: If \code{TRUE}, the local fractal dimension estimates are plotted.}

\item{LFD_Smooth_Est}{Logical: If \code{TRUE}, the smoothed estimates of local fractal dimension is plotted.}

\item{N}{Argument used for the estimation of Hurst functions. Number of sub-intervals on which the estimation is performed on. Default is set to 100 sub-intervals.}

\item{Q}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}

\item{L}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}
}
\value{
A ggplot object which plots the time series with theoretical, raw and smoothed estimates of Hurst function
and raw and smoothed estimates of local fractal dimension.
}
\description{
Creates a plot of the user provided time series with the
Hurst function estimated using \code{\link{Hurst}}, the smoothed estimated
Hurst function and local fractal dimension estimated using \code{\link{LFD}}
and smoothed estimates of local fractal dimension.
}
\examples{
TS <- data.frame("t"=seq(0,1,length=1000),"X(t)"=rnorm(1000))
plot_ts_est(TS,Raw_Est_H=TRUE,Smooth_Est_H=TRUE,LFD_Est=TRUE)

}
\seealso{
\code{\link{Hurst}}, \code{\link{LFD}}
}
