% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ex_shinyapp.R
\name{shinyapp_sim}
\alias{shinyapp_sim}
\title{Shiny app for simulation of Gaussian Haar-based multifractional processes}
\usage{
shinyapp_sim()
}
\value{
Plor
}
\description{
Plots
}
\examples{
shinyapp_sim()
}
