% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_FGN}
\alias{sim_FGN}
\title{Simulation of fractional Gaussian noise}
\usage{
sim_FGN(H, t_start, t_end, N = 1000, plot = FALSE)
}
\arguments{
\item{H}{Hurst parameter which lies between 0 and 1.}

\item{t_start}{Initial time point.}

\item{t_end}{Terminal time point.}

\item{N}{Number of time steps the interval \eqn{[t_{start},t_{end}]} is split into.
Default set to 1000.}

\item{plot}{Logical: If \code{TRUE}, the realisation of the fractional Gaussian noise
is plotted.}
}
\value{
A data frame where the first column is \code{t} and second
column is simulated values of the realisation of fractional Gaussian noise.
}
\description{
This function simulates a realisation of the fractional Gaussian noise over
time interval \eqn{[t_{start},t_{end}]} for a provided Hurst parameter.
}
\examples{
sim_FGN(H=0.3,t_start=0,t_end=1,plot=TRUE)
}
\seealso{
\code{\link{FBM}}, \code{\link{Bm}}, \code{\link{GHBMP}}
}
