% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{sojourn}
\alias{sojourn}
\title{Sojourn measure}
\usage{
sojourn(X, A, n = 10000, level = "greater", subI = NULL, plot = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence (t)
and the second the data of the time series (X(t)).}

\item{A}{Constant level as a numeric value. Checks whether X(t) greater or lower than A.}

\item{n}{Number of points time sequence to be split into. Default set to 10000.}

\item{level}{A vector of character strings which specifies whether the sojourn
measure required for X \code{"greater"} or \code{"lower"} than A. Default set to \code{"greater"}.}

\item{subI}{Sub interval of the time sequence as a vector where the lower bound
the first element and upper bound the second. Optional; if provided area represented by
the sojourn measure of the sub interval is returned. Otherwise the whole time sequence is considered.}

\item{plot}{Logical: if \code{TRUE}, the time series, constant level and the sojourn measure is plotted.}
}
\value{
Sojourn measure. If \code{plot=TRUE}, the time series, constant level and the sojourn measure is plotted.
}
\description{
Provides sojourn measure for X(t) greater or lower than the
constant level \code{A} for the provided time sequence (t) interval or a given sub interval of t.
}
\examples{
t <- seq(0,1,length=1000)
TS <- data.frame("t"=t,"X(t)"=rnorm(1000))
sojourn(TS,0.8,level='lower',subI=c(0.5,0.8),plot=TRUE)

}
