% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses_functions.R
\name{T.level}
\alias{T.level}
\title{Level crossing times}
\usage{
T.level(X, A, subI = NULL, plot = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence \eqn{(t)}
and the second the values of the time series \eqn{(X(t))}.}

\item{A}{Constant level as a numeric value.}

\item{subI}{Time sub-interval is a vector where the lower bound is
the first element and upper bound is the second. Optional: If provided level crossing
times of the sub-interval is returned, otherwise the whole time sequence is considered.}

\item{plot}{Logical: If \code{TRUE}, the time series, constant level and corresponding \eqn{t} values are plotted.}
}
\value{
The level crossing times. If \code{plot=TRUE}, a plot of the time series with
the constant level crossing and level crossing times are plotted.
}
\description{
Computes the corresponding \eqn{t} value(s) where a time series intersects a specific
constant level for the provided time interval or its sub-interval.
}
\examples{
t <- seq(0,1,length=100)
TS <- data.frame("t"=t,"X(t)"=rnorm(100))
T.level(TS,0.1,subI=c(0.2,0.8),plot=TRUE)



}
