% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_est.R
\name{H_LFD}
\alias{H_LFD}
\alias{plot.H_LFD}
\title{Plot the estimated Hurst functions and local fractal dimension estimates
for objects of class \code{"H_LFD"}}
\usage{
H_LFD(X, N = 100, Q = 2, L = 2)

\method{plot}{H_LFD}(
  x,
  Raw_Est_H = TRUE,
  Smooth_Est_H = TRUE,
  LFD_Est = TRUE,
  LFD_Smooth_Est = TRUE,
  ...
)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence \eqn{t}
and the second one is the values of the time series \eqn{X(t)}.
For accurate estimated Hurst functions, \code{X} should be of at least 500 data points.}

\item{N}{Argument used for the estimation of Hurst functions. Number of sub-intervals on which the estimation is performed on. Default is set to 100 sub-intervals.}

\item{Q}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}

\item{L}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}

\item{x}{Return from \code{\link{H_LFD}}.}

\item{Raw_Est_H}{Logical: If \code{TRUE}, the Hurst function estimated by using \code{\link{Hurst}} is plotted.}

\item{Smooth_Est_H}{Logical: If \code{TRUE}, the smoothed estimated Hurst function is plotted.
The estimated Hurst function is smoothed using the loess method.}

\item{LFD_Est}{Logical: If \code{TRUE}, the local fractal dimension estimates are plotted.}

\item{LFD_Smooth_Est}{Logical: If \code{TRUE}, the smoothed estimates of local fractal dimension is plotted.}

\item{...}{Other arguments.}
}
\value{
The return from \code{\link{H_LFD}} will be an object list of class \code{"H_LFD"} with the following components:
\describe{
\item{Raw_Hurst_estimates}{A data frame of where the first column is a time sequence and second column is estimated values of the Hurst function.}
\item{Smoothed_Hurst_estimates}{A data frame of where the first column is a time sequence and second column is smoothed estimates of the Hurst function.}
\item{LFD_estimates}{A data frame of where the first column is a time sequence and second column is Local fractal dimension estimates.}
\item{LFD_Smoothed_estimates}{A data frame of where the first column is a time sequence and second column is smoothed estimates of Local fractal dimension.}
\item{Data}{User provided time series.}}
}
\description{
Creates a plot of the user provided time series with the
Hurst function estimated using \code{\link{Hurst}}, the smoothed estimated
Hurst function and local fractal dimension estimated using \code{\link{LFD}}
for objects of class \code{"H_LFD"}.
}
\examples{
TS <- data.frame("t"=seq(0,1,length=1000),"X(t)"=rnorm(1000))
Object <- H_LFD(TS,N=100,Q=2,L=2)
plot(Object,Raw_Est_H=TRUE,Smooth_Est_H=TRUE,LFD_EST=TRUE)

}
\seealso{
\code{\link{Hurst}}, \code{\link{LFD}}, \code{\link{plot_ts_est}}
}
