% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses_functions_1.R
\name{X_min}
\alias{X_min}
\title{Estimated minimum of a time series}
\usage{
X_min(X, subI = NULL, plot = FALSE, vline = FALSE, hline = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a numeric time sequence \eqn{t}
and the second the values of the time series \eqn{X(t)}.}

\item{subI}{Time sub-interval is a vector where the lower bound is
the first element and upper bound is the second. Optional: If provided minimum of the
sub-interval is returned, otherwise the whole time interval is considered.}

\item{plot}{Logical: If \code{TRUE}, the time series, the minimum and corresponding \eqn{t} values are plotted in interactive sessions.}

\item{vline}{Logical: If \code{TRUE}, a vertical line is plotted across the minimum.}

\item{hline}{Logical: If \code{TRUE}, a horizontal line is plotted across the minimum.}
}
\value{
A list of numeric vector(s). The first element in the vector is the corresponding \eqn{t} value and second the
minimum of the time series.
}
\description{
This function computes the minimum of a time series for the provided
time interval or its sub-interval.
}
\examples{
t <- seq(0, 1, length = 100)
TS <- data.frame("t" = t, "X(t)" = rnorm(100))
X_min(TS, subI = c(0.2, 0.8), plot = TRUE)

}
\seealso{
\code{\link{X_max}}
}
