% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Bm}
\alias{Bm}
\title{Simulation of Brownian motion}
\usage{
Bm(x1 = 0, t_start = 0, t_end = 1, N = 1000, plot = FALSE)
}
\arguments{
\item{x1}{Value of the process at the initial time point.}

\item{t_start}{Initial time point.}

\item{t_end}{Terminal time point.}

\item{N}{Number of time steps the interval \eqn{[t_{start},t_{end}]} is split into.
Default set to 1000.}

\item{plot}{Logical: If \code{TRUE}, the realisation of the Brownian
motion is plotted.}
}
\value{
A data frame where the first column is \code{t} and second
column is simulated values of the realisation of Brownian motion.
}
\description{
This function simulates a realisation of the Brownian motion over
time interval \eqn{[t_{start},t_{end}]} with \code{N} increments and initial value \code{x1}.
}
\examples{
Bm(x1=0,t_start=0,t_end=2,plot=TRUE)
}
\seealso{
\code{\link{GHBMP}}, \code{\link{FBM}}, \code{\link{sim_FGN}}
}
