% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_GHBMP.R
\name{est_cov}
\alias{est_cov}
\title{Empirical covariance estimates of a process}
\usage{
est_cov(X, plot = FALSE)
}
\arguments{
\item{X}{A data frame where the first column is the time sequence and the remaining
columns are the values of each realization of the process.}

\item{plot}{Logical: If TRUE, a 3D surface plot of the covariance function is plotted.}
}
\value{
An \eqn{m \times m} matrix, where \eqn{m} is the number of time points.
Each element represents the empirical covariance estimate for the
corresponding time points which are arranged in ascending order.
}
\description{
Computes the empirical covariance estimates of a process, for each pair of time points in the
time sequence using N realizations of the process.
}
\examples{
#Matrix of empirical covariance estimates of the GHBMP with Hurst function H.
t <- seq(0,1,by=(1/2)^8)
H <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
#Only 5 realizations of GHBMP are used to reduce the computational time.
X.t <- replicate(5, GHBMP(t,H), simplify = FALSE)
X <- do.call(rbind, lapply(X.t, function(df) df[, 2]))
Data <- data.frame(t,t(X))
cov.mat <- est_cov(Data,plot=TRUE)
cov.mat

}
