% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{FBM}
\alias{FBM}
\title{Simulation of fractional Brownian motion}
\usage{
FBM(H, t_end, N = 1000, plot = FALSE)
}
\arguments{
\item{H}{Hurst parameter which lies between 0 and 1.}

\item{t_end}{Terminal time point.}

\item{N}{Number of time steps the interval \eqn{[0,t_{end}]} is split into.
Default set to 1000.}

\item{plot}{Logical: If \code{TRUE}, the realisation of the fractional Brownian
motion is plotted.}
}
\value{
A data frame where the first column is \code{t} and second
column is simulated values of the realisation of fractional Brownian motion.
}
\description{
This function simulates a realisation of the fractional Brownian motion over
time interval \eqn{[0,t_{end}]} for a provided Hurst parameter.
}
\examples{
FBM(H=0.3,t_end=1,plot=TRUE)
}
\seealso{
\code{\link{sim_FGN}}, \code{\link{Bm}}, \code{\link{GHBMP}}
}
