% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{hclust_hurst}
\alias{hclust_hurst}
\title{Hierarchical clustering of multifractional processes}
\usage{
hclust_hurst(
  X.t,
  k = NULL,
  h = NULL,
  dist.method = "euclidean",
  method = "complete",
  dendrogram = FALSE,
  N = 100,
  Q = 2,
  L = 2
)
}
\arguments{
\item{X.t}{A list of data frames where the first column is a time sequence
from 0 to 1 and the second the values of the multifractional process.
See Examples for usage.}

\item{k}{The desired number of clusters.}

\item{h}{The height where the dendrogram should be cut into. Either \code{k} or \code{h} must be specified. If both are provided k is used.}

\item{dist.method}{A string which specifies a registered distance in \link[proxy]{dist}. The default
is \code{"euclidean"}.}

\item{method}{A string which specifies the hierarchical method used. Available methods
are \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"}, \code{"mcquitty"}, \code{"median"} and \code{"centroid"}.
The default method is \code{"complete"}.}

\item{dendrogram}{Logical: If \code{TRUE} the dendrogram is plotted indicating
the clusters.}

\item{N}{Argument used for the estimation of Hurst functions. Number of sub-intervals on which the estimation is performed on. Default is set to 100 sub-intervals.}

\item{Q}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}

\item{L}{Argument used for the estimation of Hurst functions. Fixed integer greater than or equal to 2. Default is set to 2.}
}
\value{
An object list of class \code{"hc_hurst"} with print and plot methods. The list has following components:
\describe{
  \item{\code{cluster_info}}{A data frame indicating the cluster number and distance to cluster center of
  each multifractional process (Item). Distance is obtained from the \code{dist.method}.}
  \item{\code{cluster}}{A vector of cluster number of each process.}
  \item{\code{cluster_sizes}}{Number of processes in each cluster.}
  \item{\code{centers}}{A data frame of cluster centers. Center obtained as the average of each smoothed estimated Hurst function in the cluster. Columns denote each time point.
  Row names denote cluster number.}
  \item{\code{smoothed_Hurst_estimates}}{A data frame of smoothed Hurst estimates. Columns denote each time point. Rows denote estimates for each process.}
  \item{\code{raw_Hurst_estimates}}{A list of data frames of raw Hurst estimates.}
  \item{\code{call}}{The function call consisting the function applied to the provided arguments.}
}
}
\description{
This function performs hierarchical clustering of multifractional
processes based on the Hurst functions.
}
\details{
The Hurst function of each multifractional process is estimated using \code{\link{Hurst}} and the smoothed Hurst estimates are used
for the cluster analysis. The distances between each smoothed Hurst estimate is computed by the dist.method provided and passed into
the \code{\link[stats]{hclust}} for hierarchical clustering.
}
\examples{
#Simulation of multifractional processes
t <- seq(0,1,by=(1/2)^10)
H1 <- function(t) {return(0.1+0*t)}
H2 <- function(t) {return(0.2+0.45*t)}
H3 <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
X.list.1 <- replicate(3, GHBMP(t,H1),simplify = FALSE)
X.list.2 <- replicate(3, GHBMP(t,H2),simplify = FALSE)
X.list.3 <- replicate(3, GHBMP(t,H3),simplify = FALSE)
X.list <- c(X.list.1,X.list.2,X.list.3)

#Hierarchical clustering based on k=3 clusters with dendrogram plotted
HC1 <- hclust_hurst(X.list,k=3,dendrogram=TRUE)
print(HC1)

#Plot of smoothed Hurst functions in each cluster with cluster centers
plot(HC1,type ="ec")
}
\seealso{
\code{\link{print.hc_hurst}}, \code{\link{plot.hc_hurst}}, \code{\link{kmeans_hurst}}
}
