% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{Hurst}
\alias{Hurst}
\title{Statistical estimation of the Hurst function}
\usage{
Hurst(X.t, N = 100, Q = 2, L = 2)
}
\arguments{
\item{X.t}{Data frame where the first column is a time sequence from 0 to 1 and the second the data of the multifractional process.
For reliable estimates the data frame should be of at least XXX data points.}

\item{N}{Number of sub intervals the estimation is performed on. Default is set to 100 sub intervals.}

\item{Q}{Fixed integer greater than or equal to 2. Default is set to 2.}

\item{L}{Arbitrary fixed integer greater than or equal to 2. Default is set to 2.}
}
\value{
A data frame where the first column is a time sequence and second column is estimated data for the Hurst function.
Each row gives estimated values for each time point using \code{N} sub intervals.
}
\description{
This function provides statistical estimates for the Hurst function of multifractional processes.
}
\details{
Statistical estimation of the Hurst function is done based on the results of Ayache, A.,
& Bouly, F. (2023). The estimator is built through generalized quadratic variations
of the process associated with its generalized increments.
}
\examples{
#Example 1: For a multifractional process simulated using GHBMP function
T <- seq(0,1,by=(1/2)^10)
H <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
X <- GHBMP(T,H)
Hurst(X)


\dontrun{
#Example 2: For a Fractional Brownian motion simulated using the fbm function of somebm package.
library(somebm)
T <- seq(0,1,length=10000)
FBM<-data.frame("t"=T,"p"=fbm(hurst=0.7,n=9999))
Hurst(FBM)
}

}
\references{
Ayache, A. and Bouly, F. (2023). Uniformly and strongly consistent estimation for
the random Hurst function of a multifractional process. Latin American Journal of
Probability and Mathematical Statistics, 20(2):1587–1614.
}
