% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{plot.est}
\alias{plot.est}
\title{Plot real and estimated Hurst function with multifractional process}
\usage{
\method{plot}{est}(H.est, X.t, H = NULL, Raw_Est_H = TRUE, Smooth_Est_H = TRUE, ...)
}
\arguments{
\item{H.est}{Data frame of Hurst function estimated from \code{\link{Hurst}}.}

\item{X.t}{Data frame where the first column is a time sequence from 0 to 1 and the second the data of the multifractional process.
For accurate plot of the process, it should be of at least 200 data points.}

\item{H}{Theoretical Hurst function. Optional; if provided, the theoretical Hurst function is plotted.}

\item{Raw_Est_H}{Logical; if \code{TRUE}, the Hurst function estimated using \code{\link{Hurst}} is plotted.}

\item{Smooth_Est_H}{Logical; if \code{TRUE}, the smoothed estimated Hurst function is plotted.
The estimated Hurst function is smoothed using the loess method and is plotted.}

\item{...}{Other arguments}
}
\value{
Plot of the multifractional process with theoretical, estimated and smoothed Hurst functions.
}
\description{
Creates a plot of the multifractional process with theoretical Hurst function, Hurst function estimated
using \code{\link{Hurst}} and the smoothed estimated Hurst function.
}
\examples{
#Simulation of the multifractional process and estimation of the Hurst function
T <- seq(0,1,by=(1/2)^10)
H <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
X <- GHBMP(T,H)
Est_H<-Hurst(X)

#Plot of multifractional process, theoretical, estimated and smoothed Hurst functions
plot(Est_H,X,H=H)

#Plot of multifractional process, estimated and smoothed Hurst functions
plot(Est_H,X)
}
