% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_est.R
\name{plot.est}
\alias{plot.est}
\title{Plot theoretical and estimated Hurst functions}
\usage{
\method{plot}{est}(x, H = NULL, Raw_Est_H = TRUE, Smooth_Est_H = TRUE, ...)
}
\arguments{
\item{x}{Object of class `est`.}

\item{H}{Theoretical Hurst function. Optional; if provided, the theoretical Hurst function is plotted.}

\item{Raw_Est_H}{Logical; if \code{TRUE}, the Hurst function estimated using \code{\link{Hurst}} is plotted.}

\item{Smooth_Est_H}{Logical; if \code{TRUE}, the smoothed estimated Hurst function is plotted.
The estimated Hurst function is smoothed using the loess method and is plotted.}

\item{...}{Unused arguments.}
}
\value{
Plot of the theoretical, estimated and smoothed Hurst functions.
}
\description{
Creates a plot of estimated the Hurst function of class `est`with the theoretical
Hurst function and the smoothed estimated Hurst function.
}
\examples{
#Simulation of the multifractional process and estimation of the Hurst function
T <- seq(0,1,by=(1/2)^10)
H <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
X <- GHBMP(T,H)
Est_H<-Hurst(X)

#Plot of theoretical, estimated and smoothed Hurst functions
plot(Est_H,H=H)

#Plot of estimated and smoothed Hurst functions
plot(Est_H)


}
\seealso{
\code{\link{Hurst}}
}
