% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_est.R
\name{plot.mp}
\alias{plot.mp}
\title{Plot Gaussian Haar-based multifractional processes with theoretical and estimated Hurst functions}
\usage{
\method{plot}{mp}(
  x,
  H = NULL,
  Raw_Est_H = TRUE,
  Smooth_Est_H = TRUE,
  N = 100,
  Q = 2,
  L = 2,
  ...
)
}
\arguments{
\item{x}{Object of class `mp`. For accurate plot of the process, should be of at least 200 data points.}

\item{H}{Theoretical Hurst function. Optional; if provided, the theoretical Hurst function is plotted.}

\item{Raw_Est_H}{Logical; if \code{TRUE}, the Hurst function estimated using \code{\link{Hurst}} is plotted.}

\item{Smooth_Est_H}{Logical; if \code{TRUE}, the smoothed estimated Hurst function is plotted.
The estimated Hurst function is smoothed using the loess method and is plotted.}

\item{N}{Argument used for estimation of Hurst function. Number of sub intervals the estimation is performed on. Default is set to 100 sub intervals.}

\item{Q}{Argument used for estimation of Hurst function. Fixed integer greater than or equal to 2. Default is set to 2.}

\item{L}{Argument used for estimation of Hurst function. Arbitrary fixed integer greater than or equal to 2. Default is set to 2.}

\item{...}{Unused arguments.}
}
\value{
Plot of the multifractional process with theoretical, estimated and smoothed Hurst functions.
}
\description{
Creates a plot of the Gaussian Haar-based multifractional processes
of class `mp` simulated using \code{\link{GHBMP}} with theoretical Hurst function (if provided),
Hurst function estimated using \code{\link{Hurst}} (Logical) and
the smoothed estimated Hurst function (Logical).
}
\examples{
#Simulation of the multifractional process and estimation of the Hurst function
T <- seq(0,1,by=(1/2)^10)
H <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
X <- GHBMP(T,H)

#Plot of multifractional process, theoretical, estimated and smoothed Hurst functions
plot(X,H=H)

#Plot of multifractional process, estimated and smoothed Hurst functions
plot(X)


}
\seealso{
\code{\link{GHBMP}}, \code{\link{Hurst}}
}
