% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{plot.mp}
\alias{plot.mp}
\title{Title}
\usage{
\method{plot}{mp}(X.t, ...)
}
\arguments{
\item{X.t}{object of class mp}

\item{...}{other args}
}
\value{
plot
}
\description{
Title
}
\examples{
T <- seq(0,1,by=(1/2)^10)
Hurst <- function(t) {return(0.5-0.4*sin(6*3.14*t))}
Process <- GHBMP(T,Hurst)
plot(Process)

}
