% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{print.hc_hurst}
\alias{print.hc_hurst}
\title{Print method for "hc_hurst" class objects}
\usage{
\method{print}{hc_hurst}(x, ...)
}
\arguments{
\item{x}{Object from "hc_hurst" class.}

\item{...}{Additional arguments}
}
\description{
Prints the results of hierarchical clustering of multifractional processes.
}
