% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{print.hc_hurst}
\alias{print.hc_hurst}
\title{Print method for "hc_hurst" class objects}
\usage{
\method{print}{hc_hurst}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"hc_hurst"}.}

\item{...}{Unused arguments}
}
\description{
Prints the results of hierarchical clustering of realisations of processes.
}
\seealso{
\code{\link{hclust_hurst}}
}
