% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_clustering.R
\name{print.k_hurst}
\alias{print.k_hurst}
\title{Print method for "k_hurst" class objects}
\usage{
\method{print}{k_hurst}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"k_hurst"}.}

\item{...}{Other arguments.}
}
\value{
Prints an object of class \code{"k_hurst"}.
}
\description{
Prints the results of k-means clustering of realisations of processes.
}
\seealso{
\code{\link{kmeans_hurst}}
}
