% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmfrac.R
\docType{package}
\name{Rmfrac-package}
\alias{Rmfrac}
\alias{Rmfrac-package}
\title{Simulation and analyses of multifractional processes}
\description{
A collection of tools for simulating, analyzing and visualizing multifractional processes
with several other stochastic processes and user provided time series.
The package includes built-in estimation techniques for the Hurst function and Local Fractal Dimension
with plotting functions which includes customisable features
}
\section{Features}{

\itemize{
\item Simulation of Gaussian Haar-based multifractional process (GHBMP)
\item Simulation of Brownian motion, fractional Brownian motion, fractional Gaussian noise,
Brownian bridge and fractional Brownian bridge
\item Estimation of Hurst function and Local Fractal Dimension
\item Customizable plotting functions for GHBMP and user provided time series
with raw and smoothed estimates of Hurst function and Local Fractal Dimension
\item Estimation and visualisation of geometric statistics of stochastic processes
and time series: clustering based on the Hurst function, Sojourn measure, excursion area, etc.
\item An interactive Shiny application that provides options to explore and visualise
the core functionalities of the package through simulations and user-provided time series.
}
}

\seealso{
Useful links:
\itemize{
\item \url{https://github.com/Nemini-S/Rmfrac}
\item Report bugs at \url{https://github.com/Nemini-S/Rmfrac/issues}
}
}
\author{
\itemize{
\item Andriy Olenko \href{https://orcid.org/0000-0002-0917-7000}{ORCID} \href{mailto:a.olenko@latrobe.edu.au}{a.olenko@latrobe.edu.au}
\item Nemini Samarakoon \href{https://orcid.org/0009-0000-2107-1973}{ORCID} \href{mailto:neminisamarakoon95@gmail.com}{neminisamarakoon95@gmail.com}
}
}
