% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{T.level}
\alias{T.level}
\title{Level crossing times}
\usage{
T.level(X, A, subI = NULL, plot = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence (t)
and the second the data of the time series (X(t)).}

\item{A}{Constant level as a numeric value.}

\item{subI}{Sub interval of the time sequence as a vector where the lower bound
the first element and upper bound the second. Optional; if provided level crossing
times of the sub interval is provided. Otherwise the whole time sequence is considered.}

\item{plot}{Logical: if \code{TRUE}, the time series and constant level is plotted.}
}
\value{
The level crossing times. If \code{plot=TRUE}, a plot of the time series with
the constant level crossing and level crossing times are plotted.
}
\description{
Provides the corresponding t values a time series intersects a specific
constant level for the provided time sequence (t) interval or a provided sub interval of t.
}
\examples{
#Example 1
t <- seq(0,1,length=100)
TS <- data.frame("t"=t,"X(t)"=rnorm(100))
T.level(TS,0.1,subI=c(0.2,0.8),plot=TRUE)



}
