% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2.R
\name{X_min}
\alias{X_min}
\title{Minimum of a time series}
\usage{
X_min(X, subI = NULL, plot = FALSE)
}
\arguments{
\item{X}{Data frame where the first column is a time sequence (t)
and the second the data of the time series (X(t)).}

\item{subI}{Sub interval of the time sequence as a vector where the lower bound
the first element and upper bound the second. Optional; if provided minimum of the
sub interval is returned. Otherwise the whole time sequence is considered.}

\item{plot}{Logical: if \code{TRUE}, the time series and constant level is plotted.}
}
\value{
Print the minimum of the time series for a given sub interval and
the corresponding t values. If \code{plot=TRUE}, a plot of the time series with
with minimum and corresponding t values are plotted.
}
\description{
This function finds the local minimum of a time series for the provided
time sequence (t) interval or a given sub interval of t ,with corresponding t value(s).
}
\examples{
#Example 1
t <- seq(0,1,length=100)
TS <- data.frame("t"=t,"X(t)"=rnorm(100))
X_min(TS,subI=c(0.2,0.8),plot=TRUE)

}
